Work around MDEV-37478
THD::get_net_wait_timeout(): Silence GCC -Wconversion by an
explicit conversion from ulong (64 or 32 bits) to uint (32 bits).

+++ sql/sql_class.h
@@ -5748,7 +5748,7 @@ public:
         return variables.idle_transaction_timeout;
     }
 
-    return variables.net_wait_timeout;
+    return uint(variables.net_wait_timeout);
   }
 
   /**
